requirements.txt
matplotlib==3.1.2
numpy==1.17.4
opencv_python==4.5.1.48
PyTurboJPEG==1.5.3
PyYAML==6.0
setuptools==45.2.0
v4l2py==0.6.0
